chore: removed unused test file
